GFORMS-1590 - Fix usability issue of new gForm being started when inflight form has not been submitted
