+++ app/uk/gov/hmrc/gform/views/summary/summary.scala.html
@@ -42,7 +42,8 @@
     reviewerComments : Option[String],
     header: Html,
     summaryDeclaration: Html,
-    footer: Html
+    footer: Html,
+    refreshToken: Option[String]
 )(implicit request: Request[_], messages: Messages, l:LangADT, sse: SmartStringEvaluator)
 
 @heading = @{formTemplate.summarySection.title.value}
@@ -71,7 +72,7 @@
 
     @reviewerComments.map(c => alert(messages("summary.alert.title"), c))
 
-    @formWithCSRF(action = SummaryController.submit(formTemplate._id, maybeAccessCode, SummaryContinue), args =  'method -> "POST", 'class -> "js-form govuk-!-margin-bottom-9", 'autocomplete -> "off", 'novalidate -> "novalidate", 'id -> "gf-form") {
+    @formWithCSRF(action = SummaryController.submit(formTemplate._id, maybeAccessCode, SummaryContinue, refreshToken), args =  'method -> "POST", 'class -> "js-form govuk-!-margin-bottom-9", 'autocomplete -> "off", 'novalidate -> "novalidate", 'id -> "gf-form") {
 
         @if(summaryPagePurpose.forUser) {
             @CSRF.formField
@@ -90,13 +91,13 @@
             @govukButton(Button(content = Text(messages(continueLabelKey)),
                 inputType=Some("submit"),
                 value=Some("SummaryContinue"),
-                attributes = Map("formaction" -> uk.gov.hmrc.gform.gform.routes.SummaryController.submit(formTemplate._id, maybeAccessCode, SummaryContinue).path)
+                attributes = Map("formaction" -> uk.gov.hmrc.gform.gform.routes.SummaryController.submit(formTemplate._id, maybeAccessCode, SummaryContinue, refreshToken).path)
                 )
             )
             @if(renderComeBackLater) {
                 @govukButton(Button(content = Text(messages("linkText.saveAndComeBackLater")),
                     inputType=Some("submit"),
-                    attributes = Map("formaction" -> uk.gov.hmrc.gform.gform.routes.SummaryController.submit(formTemplate._id, maybeAccessCode, Exit).path),
+                    attributes = Map("formaction" -> uk.gov.hmrc.gform.gform.routes.SummaryController.submit(formTemplate._id, maybeAccessCode, Exit, refreshToken).path),
                     classes="govuk-button--secondary"
                     )
                 )
