+++ conf/app.routes
@@ -37,7 +37,7 @@ POST          /form/addtolist/remove/:formTemplateId/:maybeAccessCode/:sectionNu
 GET           /summary/pdf/:formTemplateId                                                                       uk.gov.hmrc.gform.gform.SummaryController.downloadPDF(formTemplateId: FormTemplateId, a: Option[AccessCode])
 GET           /summary/:formTemplateId/                                                                          uk.gov.hmrc.gform.gform.SummaryController.summaryById(formTemplateId: FormTemplateId, a: Option[AccessCode])
 GET           /summary/:formTemplateId/:maybeAccessCode                                                          uk.gov.hmrc.gform.gform.SummaryController.summaryById(formTemplateId: FormTemplateId, maybeAccessCode: Option[AccessCode])
-POST          /summary/:formTemplateId/:maybeAccessCode                                                          uk.gov.hmrc.gform.gform.SummaryController.submit(formTemplateId: FormTemplateId, maybeAccessCode: Option[AccessCode], action: Direction)
+POST          /summary/:formTemplateId/:maybeAccessCode                                                          uk.gov.hmrc.gform.gform.SummaryController.submit(formTemplateId: FormTemplateId, maybeAccessCode: Option[AccessCode], action: Direction, refreshToken: Option[String])
 GET           /declaration/:formTemplateId                                                                       uk.gov.hmrc.gform.gform.DeclarationController.showDeclaration(maybeAccessCode: Option[AccessCode], formTemplateId: FormTemplateId, se: SuppressErrors)
 GET           /declaration/:formTemplateId/:maybeAccessCode                                                      uk.gov.hmrc.gform.gform.DeclarationController.showDeclaration(maybeAccessCode: Option[AccessCode], formTemplateId: FormTemplateId, se: SuppressErrors)
 POST          /declaration/:formTemplateId/:maybeAccessCode                                                      uk.gov.hmrc.gform.gform.DeclarationController.submitDeclaration(formTemplateId: FormTemplateId, maybeAccessCode: Option[AccessCode], action: Direction)
